reverted debug code, moved htmlparser to utility
