+++ app/src/main/AndroidManifest.xml
@@ -11,8 +11,8 @@
     android:roundIcon="@mipmap/ic_launcher_round"
     android:supportsRtl="true"
     android:theme="@style/OppiaTheme">
-    <activity android:name=".player.exploration.ExplorationActivity">
-    <!--<activity android:name="org.oppia.app.home.HomeActivity">-->
+    <activity android:name=".player.exploration.ExplorationActivity"/>
+    <activity android:name="org.oppia.app.home.HomeActivity">
       <intent-filter>
         <action android:name="android.intent.action.MAIN"/>
         <category android:name="android.intent.category.LAUNCHER"/>
