+++ app/src/main/java/org/oppia/app/player/exploration/ExplorationActivityPresenter.kt
@@ -4,7 +4,6 @@ import androidx.appcompat.app.AppCompatActivity
 import org.oppia.app.R
 import org.oppia.app.activity.ActivityScope
 import org.oppia.app.home.HomeFragment
-import org.oppia.app.player.state.StateFragment
 import javax.inject.Inject
 
 /** The controller for [ExplorationActivity]. */
@@ -15,12 +14,12 @@ class ExplorationActivityPresenter @Inject constructor(private val activity: App
     if (getExplorationFragment() == null) {
       activity.supportFragmentManager.beginTransaction().add(
         R.id.exploration_fragment_placeholder,
-        StateFragment()
+        ExplorationFragment()
       ).commitNow()
     }
   }
 
-  private fun getExplorationFragment(): StateFragment? {
-    return activity.supportFragmentManager.findFragmentById(R.id.exploration_fragment_placeholder) as StateFragment?
+  private fun getExplorationFragment(): ExplorationFragment? {
+    return activity.supportFragmentManager.findFragmentById(R.id.exploration_fragment_placeholder) as ExplorationFragment?
   }
 }
