+++ app/src/main/java/org/oppia/app/player/state/InteractionAdapter.kt
@@ -16,12 +16,10 @@ import kotlinx.android.synthetic.main.multiple_choice_interaction_items.view.*
 import org.oppia.app.R
 import org.oppia.app.databinding.ItemSelectionInteractionItemsBinding
 import org.oppia.app.databinding.MultipleChoiceInteractionItemsBinding
+import org.oppia.util.data.HtmlParser
 import org.oppia.util.data.UrlImageParser
 
-const val CUSTOM_TAG = "oppia-noninteractive-image"
-const val HTML_TAG = "img"
-const val CUSTOM_ATTRIBUTE = "filepath-with-value"
-const val HTML_ATTRIBUTE = "src"
+
 
 const val VIEW_TYPE_MULTIPLE_CHOICE = 1
 const val VIEW_TYPE_ITEM_SELECTION = 2
@@ -97,7 +95,7 @@ class InteractionAdapter(
    internal fun bind(rawString: String?, position: Int, selectedPosition: Int) {
       binding.setVariable(BR.htmlContent, rawString)
       binding.executePendingBindings();
-      val html: Spannable = parseHtml(rawString, binding.root.tv_item_selection_contents)
+      val html: Spannable = HtmlParser(context,entity_type,entity_id).parseHtml(rawString, binding.root.tv_item_selection_contents)
       binding.root.tv_item_selection_contents.text = html
 
       binding.root.rl_checkbox_container.setOnClickListener {
@@ -115,7 +113,7 @@ class InteractionAdapter(
    internal fun bind(rawString: String?, position: Int, selectedPosition: Int) {
       binding.setVariable(BR.htmlContent, rawString)
       binding.executePendingBindings();
-      val html: Spannable = parseHtml(rawString, binding.root.tv_multiple_choice_contents)
+      val html: Spannable =  HtmlParser(context,entity_type,entity_id).parseHtml(rawString, binding.root.tv_item_selection_contents)
       binding.root.tv_multiple_choice_contents.text = html
 
       if (selectedPosition == position)
@@ -131,20 +129,5 @@ class InteractionAdapter(
     }
   }
 
-  private fun parseHtml(rawString: String?, tvContents: TextView): Spannable {
-    val html: Spannable
-    var htmlContent = rawString
-    if (htmlContent!!.contains(CUSTOM_TAG)) {
-      htmlContent = htmlContent.replace(CUSTOM_TAG, HTML_TAG, false);
-      htmlContent = htmlContent.replace(CUSTOM_ATTRIBUTE, HTML_ATTRIBUTE, false);
-      htmlContent = htmlContent.replace("&amp;quot;", "")
-    }
-    var imageGetter = UrlImageParser(tvContents, context, entity_type, entity_id)
-    if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.N) {
-      html = Html.fromHtml(htmlContent, Html.FROM_HTML_MODE_LEGACY, imageGetter, null) as Spannable
-    } else {
-      html = Html.fromHtml(htmlContent, imageGetter, null) as Spannable
-    }
-    return html
-  }
+
 }
